Refactor: Extract `desired_targets` in crate `detect-targets`
